Update example to conform to the new interfaces
