+++ examples/serviceStateManager/docker-compose.yml
@@ -1,4 +1,4 @@
-version: '3.3'
+version: '3.8'
 services:
   service-state-manager-sample-app:
     build:
@@ -7,6 +7,8 @@ services:
     ports:
       - 9000:9000
       - 8080:8080
+    environment:
+      ENABLE_EVENT_BASED_HEALTH_CHECK: "false"
     logging:
       driver: json-file
       options:
